Create file 103-fibonacci.c

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	unsigned long int i, j, k, sum;
+	int i, j, k, sum;
 
 	i = 0;
 	j = 1;
@@ -20,6 +20,6 @@ int main(void)
 		if(i % 2 == 0)
 			sum += i;
 	}
-	printf("%lu\n", sum);
+	printf("%d\n", sum);
 	return (0);
 }
